events: Rename internal module event_kinds => kinds
